requirements.txt
PyQt5
pywin32
mss
Pillow
numpy
requests